<select> update user location/ BootStrap / Location.php
User location is now selected in the <select> once the users location is known. Completed the Ajax call to location.php and reordered the flow of app.js.

Setup Bootstrap template. I have left in link <a>, Dropddown and search as i may use them later.

countryIndex.php - changed the request fron Curl to file_get_content as it quicker.