Map Layers / Home Marker / Nav Bar Population / Info Box
Added a map button that allows the user to change between map layers. Added a more traditional map and a satellite  map.

Removed the info box and put most of the information into the nav bar.
added  a click function for Home and Capital City to display and remove respective markers from the map.